Removed ADHOC_CACHE from Config
This was previously used to store a value of "novalidation", to
indicate occasions when validation should not take place - in this
case it was when a datamap did not have a type column.

I reverted to using CreateMasterUseCase class instead (there is a
CreateMasterUseCaseWithValidation to do validation). Now we have
to select the class in higher level code - we can use datamap checking
code already running to determine the correct course.