adapter now selects use case
adapter.cli now selects which use case is called based on whether the
datamap contains a "type" column. In which case, the validation version
of the use case is called - otherwise without validation. Both use case
classes are entirely separate at this point - we need to refactor both
to inherit from a super class.